% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{cv_boost_gaussian}
\alias{cv_boost_gaussian}
\title{k-fold Cross-validation for Hyperparameter Tuning for Boosting Algorithm}
\usage{
cv_boost_gaussian(
  x,
  y,
  k = 5,
  mstop = 100,
  nu = 0.1,
  seed = NULL,
  folds = NULL,
  patience = NULL
)
}
\arguments{
\item{x}{Datamatrix or Dataframe}

\item{y}{numeric response-vector}

\item{k}{number of folds}

\item{mstop}{limit to iterations}

\item{nu}{learning rate for boosting}

\item{seed}{random seed for reproducibility}

\item{folds}{optional fold vector, if random generation is not wanted}

\item{patience}{threshold for early stopping}
}
\value{
object of class "cv_boost_gaussian"
}
\description{
k-fold Cross-validation for Hyperparameter Tuning for Boosting Algorithm
}
