% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{cv_boost_grid}
\alias{cv_boost_grid}
\title{CV with Parameter Grid}
\usage{
cv_boost_grid(
  x,
  y,
  k = 5,
  mstop_grid = c(50, 100, 200),
  nu_grid = c(0.01, 0.1, 0.3),
  seed = NULL
)
}
\arguments{
\item{x}{Data in matrix or dataframe format}

\item{y}{numeric response vector}

\item{k}{number of folds}

\item{mstop_grid}{vector of mstop-values}

\item{nu_grid}{vector of nu-values}

\item{seed}{optional seed}
}
\value{
object with best parameters
}
\description{
CV with Parameter Grid
}
