% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_folds}
\alias{generate_folds}
\title{Generate CV Folds}
\usage{
generate_folds(n, k, seed = NULL)
}
\arguments{
\item{n}{number of observations}

\item{k}{number of folds}

\item{seed}{optional seed}
}
\value{
integer vector of fold assignments
}
\description{
Generate CV Folds
}
