% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{plot.cv_boost_gaussian}
\alias{plot.cv_boost_gaussian}
\title{Plot-method for CV Function "cv_boost_gaussian"}
\usage{
\method{plot}{cv_boost_gaussian}(x, ...)
}
\arguments{
\item{x}{object of class "cv_boost_gaussian"}

\item{...}{other parameters for the plot() function}
}
\description{
Plot-method for CV Function "cv_boost_gaussian"
}
