% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{print.cv_boost_gaussian}
\alias{print.cv_boost_gaussian}
\title{print-method for CV Function "cv_boost_gaussian"}
\usage{
\method{print}{cv_boost_gaussian}(x, ...)
}
\arguments{
\item{x}{object of class "cv_boost_gaussian"}

\item{...}{other parameters for the print() function}
}
\description{
print-method for CV Function "cv_boost_gaussian"
}
