% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{print.cv_boost_grid}
\alias{print.cv_boost_grid}
\title{print-method for CV Function "cv_boost_grid"}
\usage{
\method{print}{cv_boost_grid}(x, ...)
}
\arguments{
\item{x}{object of class "cv_boost_grid"}

\item{...}{other parameters for the print() function}
}
\description{
print-method for CV Function "cv_boost_grid"
}
