% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_dist_methods.R
\name{print.summary.boost_dist}
\alias{print.summary.boost_dist}
\title{Print a Distributional Boosting Model Summary}
\usage{
\method{print}{summary.boost_dist}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.boost_dist"}.}

\item{digits}{Number of digits used when printing numeric values.}

\item{...}{Further arguments, currently ignored.}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print a Distributional Boosting Model Summary
}
