% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{print.summary.cv_boost_gaussian}
\alias{print.summary.cv_boost_gaussian}
\title{extra print-method for summary-method of "cv_boost_gaussian"}
\usage{
\method{print}{summary.cv_boost_gaussian}(x, ...)
}
\arguments{
\item{x}{object of class "summary.cv_boost_gaussian"}

\item{...}{other parameters passed}
}
\description{
extra print-method for summary-method of "cv_boost_gaussian"
}
