% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{print.summary.cv_boost_grid}
\alias{print.summary.cv_boost_grid}
\title{extra print-method of summary-method of "cv_boost_grid"}
\usage{
\method{print}{summary.cv_boost_grid}(x, ...)
}
\arguments{
\item{x}{object of class "summary.cv_boost_grid"}

\item{...}{other parameters passed}
}
\description{
extra print-method of summary-method of "cv_boost_grid"
}
