% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_dist_methods.R
\name{summary.boost_dist}
\alias{summary.boost_dist}
\title{Summarize a Distributional Boosting Model}
\usage{
\method{summary}{boost_dist}(object, ...)
}
\arguments{
\item{object}{A fitted object inheriting from class \code{"boost_dist"}.}

\item{...}{Further arguments, currently ignored.}
}
\value{
An object of class \code{"summary.boost_dist"}.
}
\description{
Summarize a Distributional Boosting Model
}
