% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{summary.cv_boost_gaussian}
\alias{summary.cv_boost_gaussian}
\title{summary-method for CV Function "cv_boost_gaussian"}
\usage{
\method{summary}{cv_boost_gaussian}(object, ...)
}
\arguments{
\item{object}{of class "cv_boost_gaussian"}

\item{...}{other parameters for the summary() function}
}
\value{
list of values for model summary
}
\description{
summary-method for CV Function "cv_boost_gaussian"
}
