% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{summary.cv_boost_grid}
\alias{summary.cv_boost_grid}
\title{summary-method for CV Function "cv_boost_grid"}
\usage{
\method{summary}{cv_boost_grid}(object, ...)
}
\arguments{
\item{object}{object of class "cv_boost_grid"}

\item{...}{other parameters for the summary() function}
}
\value{
list with sorted grid results and best combination
}
\description{
summary-method for CV Function "cv_boost_grid"
}
